Missing session
After Selenium 2.34+ versions, if you run the tests, the sessionId is missing from the URL.
(BadMethodCallException: The command http://localhost:4444/wd/hub/session/url is not recognized by the server. shown)
It should be http://localhost:4444/wd/hub/session/{sessionId}/url. The sessionId is not being appended to the URL